Use getter/setter for lastPawnJump instead of direct access
I think I want to move away from having any modules directly modify
state on the board model. This is mostly heuristic, as I haven't
felt much pain around it yet, but it may help later if I try to
derive a purely read-only version of the model to provide for
querying from the view layer or elsewhere.