Simplify rule for knight movement
I'm a monster.

I knew it was perfectly readable before, that I was losing clarity in
favor of succinctness. But I just couldn't drown out the little
mathematician in my head shouting "Look, look, this expression is
equivalent in all cases because 2 is prime so the only pair of integers
for which it is the product are 1 and 2 which are the coordinate deltas
in a knight move!!!"