Block SELECT_SQUARE action during promotion
Previously this was handled inside the board model, but it seems like
something the view layer (which knows when a modal is open) should
handle? I'm not totally sold, because it is an important part of game
logic that until a pawn on the last row finishes promoting, play should
not continue. We'll have to see what happens.